Add CabinetSection
