+++ src/pages/cabinet.js
@@ -4,7 +4,7 @@ import Layout from "../components/layout"
 import SEO from "../components/seo"
 
 import StackedBarChart from "../components/stackedBarChart"
-import CabinetMemberList from "../components/cabinetMemberList"
+import CabinetMemberSection from "../components/CabinetMemberSection"
 
 // Mock for StackedBarChart
 const mpTypes = [
@@ -46,7 +46,7 @@ const CabinetPage = () => (
       <StackedBarChart data={genders}></StackedBarChart>
       <StackedBarChart data={ages}></StackedBarChart>
     </div>
-    <CabinetMemberList />
+    <CabinetMemberSection />
   </Layout>
 )
